Summary : fix issues with put/del operations, add debugging code.
1. Fixed issues with put operation on nldb_array_tree.
1.1 Prev/next links in Leaf nodes are broken.
1.2 When leaf/internal node splits, invalid key is inserted into the parent node.
1.3 When leaf/internal node splits, children nodes moved to a new node still points to the old parent.
2. Added nldb_debuggable_array_tree, which can print the structure of the tree t
o console.
3. Added consistency checking codes for nldb_array_tree.
3.1 Check prev/next link in the leaf nodes.
4. Added consistency checking codes for nldb_debuggable_array_tree.
4.1 Check keys are ordered when it prints the tree structure by visiting keys wi
th depth first search.
5. Refactored performance test codes.

+++ src/gtest/nldb_test.h
@@ -27,6 +27,7 @@
 #include <nldb/nldb_common.h>
 #include <nldb/nldb.h>
 
+#include <stdint.h>
 using namespace std;
 
 const int TEST_DB_ID = 1;
@@ -37,8 +38,8 @@ const int PERSISTENT_TABLE_ID = 2;
 #define VALUE_SIZE = 256;
 
 typedef struct tbl_key {
-	long key1;
-	long key2;
+	uint32_t key1;
+	uint32_t key2;
 } tbl_key ;
 typedef struct tbl_value{
 	long value;
@@ -96,7 +97,8 @@ extern tbl_value    var##_value; \
 extern nldb_key_t   var##_nldb_key; \
 extern nldb_value_t var##_nldb_value;
 
-
+// About using htonl : Keys are stored in big endian encoding in NLDB.
+// This is to quickly compare keys by calling memcmp on keys whenever NLDB searches a key.
 #define DECLARE_KEY_VALUE(var, k1, k2, v) \
 tbl_key      var##_key = {htonl(k1), htonl(k2)}; \
 tbl_value    var##_value = {v}; \
